refactor: Fixture 제공 객체 제거 및 함수들 root로 이동
