+++ src/test/kotlin/lotto/domain/WinningNumbersTest.kt
@@ -3,7 +3,7 @@ package lotto.domain
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
-import lotto.fixture.LottoNumbersFixtureMaker.createLottoNumbers
+import lotto.fixture.createLottoNumbers
 import lotto.fixture.of
 
 class WinningNumbersTest : FreeSpec({
